Styling of currency selector
