+++ dashboard/src/components/pricing/PricingComponent.tsx
@@ -6,6 +6,14 @@ import { PricingCards } from './PricingCards';
 import { SelectedPlan } from '@/types/pricing';
 import { EVENT_RANGES } from '@/lib/billing/plans';
 import type { Currency, UserBillingData } from '@/entities/billing';
+import {
+  Select,
+  SelectContent,
+  SelectGroup,
+  SelectItem,
+  SelectTrigger,
+  SelectValue,
+} from '@/components/ui/select';
 
 interface PricingComponentProps {
   onPlanSelect?: Dispatch<SelectedPlan>;
@@ -23,6 +31,7 @@ export function PricingComponent({
   defaultCurrency = 'USD',
 }: PricingComponentProps) {
   const [selectedRangeIndex, setSelectedRangeIndex] = useState(initialRangeIndex);
+  const [selectedCurrency, setSelectedCurrency] = useState<Currency>(defaultCurrency);
   const currentRange = EVENT_RANGES[selectedRangeIndex];
 
   const handleSliderChange = useCallback((e: React.ChangeEvent<HTMLInputElement>) => {
@@ -30,15 +39,32 @@ export function PricingComponent({
     setSelectedRangeIndex(index);
   }, []);
 
+  const handleCurrencyChange = useCallback((currency: string) => {
+    setSelectedCurrency(currency as Currency);
+  }, []);
+
   return (
     <div className={className}>
-      <div className='mb-12 flex items-center justify-center gap-8'>
-        <div className='mx-1 w-full lg:w-1/2'>
+      <div className='mx-1 mb-8 grid grid-cols-5 items-end justify-center gap-6'>
+        <div className='col-span-5 col-start-1 lg:col-span-3 lg:col-start-2'>
           <PricingSlider
             currentRange={currentRange}
             selectedRangeIndex={selectedRangeIndex}
             handleSliderChange={handleSliderChange}
           />
+        </div>
+        <div className='text-muted-foreground col-start-3 flex flex-shrink-0 justify-center text-xs lg:col-start-5 lg:justify-end'>
+          <Select value={selectedCurrency} onValueChange={handleCurrencyChange}>
+            <SelectTrigger size='sm'>
+              <SelectValue />
+            </SelectTrigger>
+            <SelectContent>
+              <SelectGroup>
+                <SelectItem value='USD'>USD ($)</SelectItem>
+                <SelectItem value='EUR'>EUR (€)</SelectItem>
+              </SelectGroup>
+            </SelectContent>
+          </Select>
         </div>
       </div>
 
@@ -47,7 +73,7 @@ export function PricingComponent({
         onPlanSelect={onPlanSelect}
         mode={onPlanSelect ? 'billing' : 'landing'}
         billingData={billingData}
-        currency={defaultCurrency}
+        currency={selectedCurrency}
       />
     </div>
   );
